Use place_id instead of id as per:
https://developers.google.com/places/documentation/search
"
Note: The id and reference fields are deprecated as of June 24, 2014.
They are replaced by the new place ID, a unique identifier that can be
used to compare places and to retrieve information about a place.
The Places API currently returns a place_id in all responses,
and accepts a placeid in the Place Details and Place Delete requests.
Soon after June 24, 2015, the API will stop returning the id and
reference fields in responses. Some time later, the API will no
longer accept the reference in requests.
We recommend that you update your code to use the new place ID
instead of id and reference as soon as possible."